completed adding Trips admin page

+++ src/main/java/com/sda/projects/Travel_agency/repos/AirportRepository.java
@@ -3,5 +3,5 @@ package com.sda.projects.Travel_agency.repos;
 import com.sda.projects.Travel_agency.entity.Airport;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AirportRepository extends CrudRepository<Airport, Long> {
+public interface AirportRepository extends CrudRepository<Airport, Integer> {
 }
