+++ src/main/java/com/sda/projects/Travel_agency/repos/CityRepository.java
@@ -3,5 +3,5 @@ package com.sda.projects.Travel_agency.repos;
 import com.sda.projects.Travel_agency.entity.City;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CityRepository extends CrudRepository<City, Long> {
+public interface CityRepository extends CrudRepository<City, Integer> {
 }
