+++ src/main/java/com/sda/projects/Travel_agency/repos/ContinentRepository.java
@@ -3,5 +3,5 @@ package com.sda.projects.Travel_agency.repos;
 import com.sda.projects.Travel_agency.entity.Continent;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ContinentRepository extends CrudRepository<Continent, Long> {
+public interface ContinentRepository extends CrudRepository<Continent, Integer> {
 }
