+++ src/main/java/com/sda/projects/Travel_agency/repos/CountryRepository.java
@@ -3,5 +3,5 @@ package com.sda.projects.Travel_agency.repos;
 import com.sda.projects.Travel_agency.entity.Country;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CountryRepository extends CrudRepository<Country, Long> {
+public interface CountryRepository extends CrudRepository<Country, Integer> {
 }
