+++ src/main/java/com/sda/projects/Travel_agency/repos/HotelRepository.java
@@ -3,5 +3,5 @@ package com.sda.projects.Travel_agency.repos;
 import com.sda.projects.Travel_agency.entity.Hotel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HotelRepository extends CrudRepository<Hotel, Long> {
+public interface HotelRepository extends CrudRepository<Hotel, Integer> {
 }
