+++ src/main/java/com/sda/projects/Travel_agency/repos/TripRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface TripRepository extends CrudRepository<Trip, Long> {
+public interface TripRepository extends CrudRepository<Trip, Integer> {
 }
